requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-Login>=0.6.0
Authlib>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0